Extract CosmWasmClient.searchTx
